I've slightly modified the definition of actions as follows:
An action (as they should be given in the column specification)
consist now of a <Notification> and a <default value>:

* The <Notification> can be either "None", "Warning" or "Error" ---as
  before. If no <Notification> is specified, then "None" is assumed by
  default.

* The <default value> can be any one given between parenthesis only
  if the <Notification> is given before, e.g., "Warning (3.14159)",
  but the parenthesis can be omitted if and only if no <Notification>
  is given, e.g., "3.14159". Of course, if the <Notification> is
  "Error", then no <default value> is expected (as before) because the
  program immediately aborts execution.

An important consequence of this redefinitions is that if a column has
no default value and no value is found in a specific cell then it is
plainly ignored and no data is written to the database. This is cool!
If you want some info to be written to the database when data is no
found, use <default value> but if you want to filter data, then give
no <default value>

This way, it is now possible to create different specifications as
shown in the following table:

test_table_2 using "examples/spsheets/alumnos-ene.ods":"Calificaciones" {
	   nombre    $B3:$B66 text Warning;
	   apellidos $C3:$C66 text Error;
	   practicas $O3:$O66 real Warning (0);
	   teoria    $P3:$P66 real 0;
	   examen    $Q3:$Q66 real;
}

Note that the column "nombre" specifies no <default value> so that if
any of the cells in the range $B3:$B66 is empty, then the whole record
is ignored and no data is written to the database; "teoria" is
specified with the default <Notification> "None" and thus, the
<default value> can be specified without parenthesis; "examen" has
neither a <Notification> nor a <default value>, hence it is a shortcut
for "None (0)"

Also, when accesing the spreadsheet if an error is issued by pyexcel,
this is captured and data is assumed just to be empty. This allows the
user to specify any ranges in the spreadsheet even if they go beyond
the bounds.