This version provides full support to dragging formulae.
However, one the spreadsheet is created, some cells might contain an
error. The solution is simple: to take any, delete any character of
the formula and to re-insert it, forcing the evaluation of the formula
this way. Next, drag manually. This is very unfortunate and I'd like
to know if there is a way to do it for all cells at the same time as
all formulae are correctly computed